Add SignedNumber and SignedInteger extensions
