+++ Sources/Extensions/Foundation/FloatingPointExtensions.swift
@@ -12,19 +12,6 @@ import Foundation
 // MARK: - Properties
 public extension FloatingPoint {
 	
-	/// SwifterSwift: Absolute of number.
-	public var abs: Self {
-		return Swift.abs(self)
-	}
-	
-	/// SwifterSwift: String with number and current locale currency.
-	public var asLocaleCurrency: String {
-		let formatter = NumberFormatter()
-		formatter.numberStyle = .currency
-		formatter.locale = Locale.current
-		return formatter.string(from: self as! NSNumber)!
-	}
-	
 	/// SwifterSwift: Ceil of number.
 	public var ceil: Self {
 		return Foundation.ceil(self)
@@ -38,21 +25,6 @@ public extension FloatingPoint {
 	/// SwifterSwift: Floor of number.
 	public var floor: Self {
 		return Foundation.floor(self)
-	}
-	
-	/// SwifterSwift: Check if number is positive.
-	public var isPositive: Bool {
-		return self > 0
-	}
-	
-	/// SwifterSwift: Check if number is negative.
-	public var isNegative: Bool {
-		return self < 0
-	}
-	
-	/// SwifterSwift: String.
-	public var string: String {
-		return String(describing: self as! NSNumber)
 	}
 	
 	/// SwifterSwift: Degree value of radian input.
