+++ Sources/Extensions/Foundation/IntExtensions.swift
@@ -6,28 +6,12 @@
 //  Copyright © 2016 Omar Albeik. All rights reserved.
 //
 
-#if os(macOS)
-	import Cocoa
-#else
-	import UIKit
-#endif
+import Foundation
 
 
 // MARK: - Properties
 public extension Int {
 	
-	/// SwifterSwift: Absolute value of integer.
-	public var abs: Int {
-		return Swift.abs(self)
-	}
-	
-	/// SwifterSwift: String with number and current locale currency.
-	public var asLocaleCurrency: String {
-		let formatter = NumberFormatter()
-		formatter.numberStyle = .currency
-		formatter.locale = Locale.current
-		return formatter.string(from: self as NSNumber)!
-	}
 	
 	/// SwifterSwift: CountableRange 0..<Int.
 	public var countableRange: CountableRange<Int> {
@@ -39,15 +23,9 @@ public extension Int {
 		return Double.pi * Double(self) / 180.0
 	}
 	
-	/// SwifterSwift: Array of digits of integer value.
-	public var digits: [Int] {
-		var digits: [Int] = []
-		for char in String(self).characters {
-			if let int = Int(String(char)) {
-				digits.append(int)
-			}
-		}
-		return digits
+	/// SwifterSwift: Degree value of radian input
+	public var radiansToDegrees: Double {
+		return Double(self) * 180 / Double.pi
 	}
 	
 	/// SwifterSwift: Number of digits of integer value.
@@ -55,25 +33,6 @@ public extension Int {
 		return String(self).characters.count
 	}
 	
-	/// SwifterSwift: Check if integer is even.
-	public var isEven: Bool {
-		return (self % 2) == 0
-	}
-	
-	/// SwifterSwift: Check if integer is odd.
-	public var isOdd: Bool {
-		return (self % 2) != 0
-	}
-	
-	/// SwifterSwift: Check if integer is positive.
-	public var isPositive: Bool {
-		return self > 0
-	}
-	
-	/// SwifterSwift: Check if integer is negative.
-	public var isNegative: Bool {
-		return self < 0
-	}
 	
 	/// SwifterSwift: UInt.
 	public var uInt: UInt {
@@ -93,16 +52,6 @@ public extension Int {
 	/// SwifterSwift: CGFloat.
 	public var cgFloat: CGFloat {
 		return CGFloat(self)
-	}
-	
-	/// SwifterSwift: String.
-	public var string: String {
-		return String(self)
-	}
-	
-	/// SwifterSwift: Degree value of radian input
-	public var radiansToDegrees: Double {
-		return Double(self) * 180 / Double.pi
 	}
 	
 	/// SwifterSwift: Roman numeral string from integer (if applicable).
@@ -130,26 +79,6 @@ public extension Int {
 		return romanValue
 	}
 	
-	/// SwifterSwift: String of format (XXh XXm) from seconds Int.
-	public var timeString: String {
-		guard self > 0 else {
-			return "0 sec"
-		}
-		if self < 60 {
-			return "\(self) sec"
-		}
-		if self < 3600 {
-			return "\(self / 60) min"
-		}
-		let hours = self / 3600
-		let mins = (self % 3600) / 60
-		
-		if hours != 0 && mins == 0 {
-			return "\(hours)h"
-		}
-		return "\(hours)h \(mins)m"
-	}
-	
 	/// SwifterSwift: String formatted for values over ±1000 (example: 1k, -2k, 100k, 1kk, -5kk..)
 	public var kFormatted: String {
 		var sign: String {
@@ -171,22 +100,6 @@ public extension Int {
 
 // MARK: - Methods
 public extension Int {
-	
-	/// SwifterSwift: Greatest common divisor of integer value and n.
-	///
-	/// - Parameter n: integer value to find gcd with.
-	/// - Returns: greatest common divisor of self and n.
-	public func gcd(of n: Int) -> Int {
-		return n == 0 ? self : n.gcd(of: self % n)
-	}
-	
-	/// SwifterSwift: Least common multiple of integer and n.
-	///
-	/// - Parameter n: integer value to find lcm with.
-	/// - Returns: least common multiple of self and n.
-	public func lcm(of n: Int) -> Int {
-		return (self * n).abs / gcd(of: n)
-	}
 	
 	/// SwifterSwift: Random integer between two integer values.
 	///
